[ObjC] Defer to the LLVM backend for unaligned atomic load and stores
LLVM can codegen correctly for atomics of any alignment in IR, on any architecture (via using libcalls when necessary).

So, the ObjC code's hasUnalignedAtomics can be removed. We can use copyStruct if the size is too big.

+++ clang/test/CodeGenObjC/objc_copyStruct.m
@@ -2,7 +2,7 @@
 // RUN: %clang -target x86_64-apple-ios -fobjc-runtime=ios -Wno-objc-root-class -S -o - -emit-llvm %s | FileCheck %s
 
 struct S {
-  float f, g;
+  double f, g;
 };
 
 @interface I
@@ -13,4 +13,3 @@ struct S {
 @end
 
 // CHECK: declare {{.*}}void @objc_copyStruct(ptr, ptr, i64, i1, i1)
-
